requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Werkzeug==2.3.7
nltk==3.8.1
numpy==1.25.2
networkx==3.1
python-docx==1.0.1
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.0.0
SpeechRecognition==3.10.0
pydub==0.25.1
python-dotenv==1.0.0
gunicorn==21.2.0